fix: fix error when merge with 3.30 assignment
