+++ Assets/Scripts/3.28/Enemy.cs
@@ -4,62 +4,83 @@ using UnityEngine;
 
 public class Enemy : Character
 {
-    // 1. _player 변수 삭제 -> GetCharacter로 접근할 거임
-    //private Player _player;
-    private float _randomHeal;
+    private Player _player;
 
+    private int DAMAGE_INCREASE = 3;
+    private int FINAL_ROUND = 10;
+    private int PLAYER_HP = 100;
+    private int SPECIAL_HEAL_MAX_VALUE = 3;
+    private int HEAL_INCREASE = 10;
+
+    /// <summary>
+    /// 1. Init: 초기화 기능
+    /// 1) Subject에 Observer로 등록
+    /// 2) _myName, _myHp, _myDamage 초기화
+    /// 3) _myName은 무조건 "Enemy"로 할 것
+    /// 4) _myHp, _myDamage는 100, 10으로 각각 초기화 (권장 사항)
+    /// </summary>
     protected override void Init()
     {
         base.Init();
         _myName = "Enemy";
-        _myHpMax = 100;
-        _myHp = _myHpMax;
-        _myDamage = 15;
-        GameManager.Instance().AddCharacter(this.GetComponent<Enemy>());
-    }
-
-    private void Awake()
-    {
-        Init();
-    }
-
-    public override void Attack()
-    {
-        if(_myName.Equals(_whoseTurn) && !_isFinished)
-        {
-            _myDamage += 3;
-            // 1. GetCharacter로 Player 접근
-            if (_gameRound >= 10) _myDamage = GameManager.Instance().GetCharacter("Player")._myHp;
-            AttackMotion();
-            GameManager.Instance().GetCharacter("Player").GetHit(_myDamage);
-        }
-    }
-
-    public override void GetHit(float damage)
-    {
-        base.GetHit(damage);
-        if(_myHp > 0)
-        {
-            _randomHeal = Random.Range(0, 10);
-            if(_randomHeal < 5) // 50% 확률로 Heal
-            {
-                StartCoroutine(HealCoroutine());
-            }
-        }
+        _myHp = 100;
+        _myDamage = 10;
     }
 
     /// <summary>
-    /// HealCoroutine: 
-    /// 1) Player가 Enemy 공격 -> Hp 깎임 -> UI 반영
-    /// 2) Enemy 확률적으로 회복 -> Hp 참 -> UI 반영
-    /// 3) 중간에 yield return 하지 않으면 한번에 처리돼서 피격 하고 Heal 하는 UI 반영이 제대로 이루어지지 않음.
+    /// 1) _player가 할당이 안됐다면,
+    /// 2) GameObject.FindWithTag 이용해서 _player 할당
     /// </summary>
-    /// <returns></returns>
-    IEnumerator HealCoroutine()
+    private void Awake()
     {
-        yield return new WaitForSeconds(1.3f);
-        _myHp += 10;
-        Debug.Log($"{_myName} Heal!");
+        Init();
+        if (_player == null)
+        {
+            _player = GameObject.FindWithTag("Player").GetComponent<Player>();
+        }
+    }
+
+    
+ 
+    /// <summary>
+    /// Attack:
+    /// 1) _gameRound가 지날때마다 데미지 3씩 증가
+    /// 2) _gameRound가 10이 되면 무조건 Player를 죽이도록 데미지 증가
+    /// </summary>
+    public override void Attack()
+    {
+        if (!_isFinished && _myName == _whoseTurn)
+        {
+            AttackMotion();
+            if (_gameRound == FINAL_ROUND)
+            {
+                _myDamage = PLAYER_HP;
+            } else
+            {
+                _myDamage = _myDamage + DAMAGE_INCREASE;
+            }
+            _player.GetHit(_myDamage);
+        }
+
+        
+    }
+
+    /// <summary>
+    /// GetHit:
+    /// 1) Player의 _randomAttack과 동일한 기능
+    /// 2) 30%의 확률로 피격시 10 체력 증가
+    ///   + Debug.Log($"{_myName} Heal!"); 추가
+    /// </summary>
+    public override void GetHit(float damage)
+    {
+        base.GetHit(damage);
+        int _randomHealOrNot = Random.Range(0, 10);
+
+        if(0 <= _randomHealOrNot && _randomHealOrNot < SPECIAL_HEAL_MAX_VALUE)
+        {
+            PLAYER_HP += HEAL_INCREASE;
+            Debug.Log($"{_myName} Heal!");
+        }
     }
 }
