+++ Assets/Scripts/3.28/Enemy.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 702a6e5417576e245a9a91e717f6e521
+guid: a4157e4f3b9612e4596848e49d48ed0e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
