+++ Assets/Scripts/3.28/GameManager.cs
@@ -4,84 +4,97 @@ using UnityEngine;
 
 public class GameManager : MonoBehaviour, Subject
 {
+    // 1. Singleton Pattern: Instance() method
+    
     private static GameManager _instance;
+
     public static GameManager Instance()
     {
         return _instance;
     }
 
-    private void Awake()
-    {
-        if (_instance == null)
-        {
-            _instance = this;
-            DontDestroyOnLoad(this.gameObject);
-        }
-        else
-        {
-            if (this != _instance)
-            {
-                Destroy(this.gameObject);
-            }
-        }
-    }
-
+    // 초기화 설정 바꾸지 말 것
     private int _gameRound = 0;
     private string _whoseTurn = "Enemy";
     private bool _isEnd = false;
 
-    // 1. SceneUI가 GameManager 접근 할 수 있도록 캐릭터 딕셔너리 선언
-    private Dictionary<string, Character> _characterList = new Dictionary<string, Character>();
+    // delegate: TurnHandler, FinishHandler 선언
+    delegate void TurnHandler(int round, string turn);
+    delegate void FinishHandler(bool isFinish);
 
-    private delegate void TurnHandler(int round, string turn);
-    private TurnHandler _turnHandler;
-    private delegate void FinishHandler(bool isFinish);
-    private FinishHandler _finishHandler;
-    // 2. UIHandler 선언 (이번에는 round, turn, isFinish 모두 받는다)
-    private delegate void UIHandler(int round, string turn, bool isFinish);
-    private UIHandler _uiHandler;
+    TurnHandler _turnHandler;
+    FinishHandler _finishHandler;
 
+    void Start()
+    {
+      
+        
+    }
+    /// <summary>
+    /// 2. RoundNotify:
+    /// 1) 현재 턴이 Enemy이면 다음 gameRound로
+    ///  + Debug.Log($"GameManager: Round {gameRound}.");
+    /// 2) TurnNotify() 호출
+    /// </summary>
     public void RoundNotify()
     {
-        if (!_isEnd)
+        if(_whoseTurn == "Enemy")
         {
-            if(_whoseTurn == "Enemy")
-            {
-                _gameRound++;
-                Debug.Log($"GameManager: Round {_gameRound}.");
-            }
-            TurnNotify();
+            _gameRound += 1;
+            Debug.Log($"GameManager: Round {_gameRound}.");
         }
+        TurnNotify();
     }
 
+    /// <summary>
+    /// 3. TurnNotify:
+    /// 1) whoseTurn update
+    ///  + Debug.Log($"GameManager: {_whoseTurn} turn.");
+    /// 2) _turnHandler 호출
+    /// </summary>
     public void TurnNotify()
     {
-        _whoseTurn = _whoseTurn == "Enemy" ? "Player" : "Enemy";
-        Debug.Log($"GameManager: {_whoseTurn} turn.");
+       if(_whoseTurn == "Enemy")
+        {
+            _whoseTurn = "Player";
+        } else
+        {
+            _whoseTurn = "Enemy";
+        }
+        
         _turnHandler(_gameRound, _whoseTurn);
-        // 2. _uiHandler 호출
+        Debug.Log($"GameManager: {_whoseTurn} turn.");
+
+        
     }
 
+    /// <summary>
+    /// 4. EndNotify: 
+    /// 1) isEnd update
+    ///  + Debug.Log("GameManager: The End");
+    ///  + Debug.Log($"GameManager: {_whoseTurn} is Win!");
+    /// 2) _finishHandler 호출
+    /// </summary>
     public void EndNotify()
     {
         _isEnd = true;
-        _finishHandler(_isEnd);
-        // 2. _uiHandler 호출
         Debug.Log("GameManager: The End");
         Debug.Log($"GameManager: {_whoseTurn} is Win!");
+        _finishHandler(_isEnd);
+
     }
 
+    // 5. AddCharacter: _turnHandler, _finishHandler 각각에 메소드 추가
     public void AddCharacter(Character character)
     {
-        _turnHandler += new TurnHandler(character.TurnUpdate);
-        _finishHandler += new FinishHandler(character.FinishUpdate);
-        // 1. _characterList에 추가
+        _turnHandler += new TurnHandler(character.GetComponent<Character>().TurnUpdate);
+        _finishHandler += new FinishHandler(character.GetComponent<Character>().FinishUpdate);
     }
 
     // 3. AddUI: SceneUI 옵저버로 등록
     public void AddUI(SceneUI ui)
     {
-    
+
     }
 
     /// <summary>
