+++ Assets/Scripts/3.28/Player.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8b73c7d95a2a04440b66cebee09f400e
+guid: 784efd2c170a4374b8d0dc220283075c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
